requirements.txt
python-telegram-bot==20.3
pandas==2.0.3
sqlalchemy==2.0.23
selenium==4.15.2
webdriver-manager==3.8.6
python-dotenv==1.0.0
openpyxl==3.1.2
lxml==4.9.2